P7

+++ java-project-healthcalc/src/main/java/healthcalc/Person.java
@@ -1,8 +1,8 @@
 package healthcalc;
 
 public interface Person {
-    float weight();
-    float height();
+    double weight();
+    double height();
     Gender gender();
     int age();
 }
